Attempt to stop constant semantic parsing
Misc. changes to try and stop semantic from parsing all the time.  I
do not recall which of the changes here are necessary for the parsing
to stop.  Perhaps all of the changes here are absolutely necessary,
perhaps not.  I do seem to recall that company-capf might have been
one of the critical ones, but aside from that, I do not know.

By extension, this also means that some of the changes here MIGHT be
detrimental and have larger effects than just stopping the constant
re-parsing of header files...  If I do find that I'm experiencing
troubles, I will need to look further into this...