Rewrite configuration, using use-package and cleaning up custom files
This is the first attempt at cleaning up my emacs configuration.  I
have tried, where possible, to get rid of all config code from the
custom/setup-*.el files -- I've found this overly difficult to manage,
since it is often unclear to me which file should be used for a given
package.  This is true both when adding new packages and when trying
to find existing config information.

Instead, wherever possible, all of the config has been revised to make
use of the use-package system, with all of the relevant configuration
being placed directly in init.el.

It is my expectation that code folding can and will be used for this
file, so that the file can be browsed easily with only the
'use-package XXX' line appearing, and the rest of the config for the
package available when unfolded.

For now, there is still a bit of config settings in some of the old
setup-*.el files.  These should eventually be migrated over.  However,
I have managed to remove a number of the files from the config (for
now, they are commented-out so they are not access by the require function)

Additionally, I have annotated the new init.el file in a few places
with ;; TODO comments, to identify parts of the config which need
further attention.

It is my hope and intent that this should provide a usable config,
allowing emacs to be fully utilized right now and eliminating the
urgency in finishing to migrate the rest of the old config.